commit 20/05/2025 venta estadistica falta terminar

+++ src/main/java/com/proyectoSGV/demo/main/ProductosExistencias.java
@@ -15,90 +15,73 @@ public class ProductosExistencias {
 	@Column(name = "idproductos")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private int idProductos;
-	
+
 	@Column(name = "codigo")
-	private int codigo;
-	
+	private long codigo;
+
 	@Column(name = "cantidad")
 	private int cantidad;
-	
+
 	@Column(name = "nomproducto")
 	private String nomProducto;
-	
-	
+
 	@Column(name = "preciocosto")
 	private double precioCosto;
 
 	@Column(name = "precioventa")
 	private double precioVenta;
 
-	
-	
-	
 	public ProductosExistencias() {
 		super();
 	}
-
 
 	public int getIdProductos() {
 		return idProductos;
 	}
 
-
 	public void setIdProductos(int idProductos) {
 		this.idProductos = idProductos;
 	}
 
-
-	public int getCodigo() {
+	public long getCodigo() {
 		return codigo;
 	}
 
-
-	public void setCodigo(int codigo) {
+	public void setCodigo(long codigo) {
 		this.codigo = codigo;
 	}
-
 
 	public int getCantidad() {
 		return cantidad;
 	}
 
-
 	public void setCantidad(int cantidad) {
 		this.cantidad = cantidad;
 	}
-
 
 	public String getNomProducto() {
 		return nomProducto;
 	}
 
-
 	public void setNomProducto(String nomProducto) {
 		this.nomProducto = nomProducto;
 	}
-
 
 	public double getPrecioCosto() {
 		return precioCosto;
 	}
 
-
 	public void setPrecioCosto(double precioCosto) {
 		this.precioCosto = precioCosto;
 	}
-
 
 	public double getPrecioVenta() {
 		return precioVenta;
 	}
 
-
 	public void setPrecioVenta(double precioVenta) {
 		this.precioVenta = precioVenta;
 	}
-
 
 	@Override
 	public String toString() {
@@ -107,11 +90,4 @@ public class ProductosExistencias {
 				+ "]";
 	}
 
-
-	
-
-	
-	
-	
-	
 }
